refactor: tag entity 분리
