+++ src/entity/tag/model/types.ts
@@ -1,0 +1,5 @@
+export interface Tag {
+  name: string
+  slug: string
+  url: string
+}
