+++ src/entity/tag/model/useTags.ts
@@ -1,0 +1,10 @@
+import { useQuery } from "@tanstack/react-query"
+
+import { getTagList } from "@/entity/tag/api/getTagList"
+
+export const useTags = () => {
+  return useQuery({
+    queryKey: ["tags"],
+    queryFn: getTagList,
+  })
+}
